[Infra] Add Windows build support
This patch adds rudimentary windows build support. A complete
platform implementation is not included, so not all targets
are added to `gn_all` yet.

Bug: 236163082

+++ osp/impl/presentation/presentation_receiver.cc
@@ -116,7 +116,7 @@ ErrorOr<size_t> Receiver::OnStreamMessage(uint64_t instance_id,
                    "kPresentationUrlAvailabilityRequest");
       OSP_VLOG << "got presentation-url-availability-request";
       msgs::PresentationUrlAvailabilityRequest request;
-      ssize_t decode_result = msgs::DecodePresentationUrlAvailabilityRequest(
+      int64_t decode_result = msgs::DecodePresentationUrlAvailabilityRequest(
           buffer, buffer_size, request);
       if (decode_result < 0) {
         if (decode_result == msgs::kParserEOF) {
@@ -143,7 +143,7 @@ ErrorOr<size_t> Receiver::OnStreamMessage(uint64_t instance_id,
       TRACE_SCOPED(TraceCategory::kPresentation, "kPresentationStartRequest");
       OSP_VLOG << "got presentation-start-request";
       msgs::PresentationStartRequest request;
-      const ssize_t result =
+      const int64_t result =
           msgs::DecodePresentationStartRequest(buffer, buffer_size, request);
       if (result < 0) {
         if (result == msgs::kParserEOF) {
@@ -210,7 +210,7 @@ ErrorOr<size_t> Receiver::OnStreamMessage(uint64_t instance_id,
                    "kPresentationConnectionOpenRequest");
       OSP_VLOG << "Got a presentation-connection-open-request";
       msgs::PresentationConnectionOpenRequest request;
-      const ssize_t result = msgs::DecodePresentationConnectionOpenRequest(
+      const int64_t result = msgs::DecodePresentationConnectionOpenRequest(
           buffer, buffer_size, request);
       if (result < 0) {
         if (result == msgs::kParserEOF) {
@@ -288,7 +288,7 @@ ErrorOr<size_t> Receiver::OnStreamMessage(uint64_t instance_id,
                    "kPresentationTerminationRequest");
       OSP_VLOG << "got presentation-termination-request";
       msgs::PresentationTerminationRequest request;
-      const ssize_t result = msgs::DecodePresentationTerminationRequest(
+      const int64_t result = msgs::DecodePresentationTerminationRequest(
           buffer, buffer_size, request);
       if (result < 0) {
         if (result == msgs::kParserEOF) {
