+++ osp/impl/presentation/url_availability_requester.cc
@@ -450,7 +450,7 @@ ErrorOr<size_t> UrlAvailabilityRequester::ReceiverRequester::OnStreamMessage(
   switch (message_type) {
     case msgs::Type::kPresentationUrlAvailabilityResponse: {
       msgs::PresentationUrlAvailabilityResponse response;
-      ssize_t result = msgs::DecodePresentationUrlAvailabilityResponse(
+      int64_t result = msgs::DecodePresentationUrlAvailabilityResponse(
           buffer, buffer_size, response);
       if (result < 0) {
         if (result == msgs::kParserEOF) {
@@ -484,7 +484,7 @@ ErrorOr<size_t> UrlAvailabilityRequester::ReceiverRequester::OnStreamMessage(
     }
     case msgs::Type::kPresentationUrlAvailabilityEvent: {
       msgs::PresentationUrlAvailabilityEvent event;
-      ssize_t result = msgs::DecodePresentationUrlAvailabilityEvent(
+      int64_t result = msgs::DecodePresentationUrlAvailabilityEvent(
           buffer, buffer_size, event);
       if (result < 0) {
         if (result == msgs::kParserEOF) {
